Move `assert()` function used in feed-parser into a base component
